Inferring trajectories with cellTree

Arguments passed to this function will be used as default parameters for the method.

description_ctmaptpx(method = "maptpx", sd_filter = 0.5,
  absolute_width = 0, width_scale_factor = 1.5,
  outlier_tolerance_factor = 0.1, rooting_method = "null",
  num_topics_lower = 2L, num_topics_upper = 15L, tot_iter = 1e+06,
  tolerance = 0.05)

description_ctgibbs(method = "Gibbs", sd_filter = 0.5, absolute_width = 0,
  width_scale_factor = 1.5, outlier_tolerance_factor = 0.1,
  rooting_method = "null", num_topics = 4L, tot_iter = 200,
  tolerance = 1e-05)

description_ctvem(method = "VEM", sd_filter = 0.5, absolute_width = 0,
  width_scale_factor = 1.5, outlier_tolerance_factor = 0.1,
  rooting_method = "null", num_topics = 4L, tot_iter = 1e+06,
  tolerance = 1e-05)

Arguments

sd_filter
Standard-deviation threshold below which genes should be removed from the data.

absolute_width
Distance threshold below which a cell vertex is considered to be attached to a backbone vertex (see paper for more details). By default, this threshold is computed dynamically, based on the distance distribution for each branch.

width_scale_factor
A scaling factor for the dynamically-computed distance threshold (ignored if absolute_width is provided). Higher values will result in less branches in the backbone tree, while lower values might lead to a large number of backbone branches.

rooting_method
Method used to root the backbone tree. Must be one of: ‘null’, ‘longest.path’, ‘center.start.group’ or ‘average.start.group’. ‘longest.path' picks one end of the longest shortest-path between two vertices. 'center.start.group’ picks the vertex in the starting group with lowest mean-square-distance to the others. ‘average.start.group’ creates a new artificial vertex, as the average of all cells in the starting group. ‘null’ picks the best method based on the type of grouping and start group information available.

num_topics_lower
The lower bound of topics to be fitted in the model.

num_topics_upper
The upper bound of topics to be fitted in the model.

tot_iter
Number of iterations of the LDA inference.

tolerance
Tolerance values of the LDA inference.

num_topics
The number of topics to fit in the model.

outlier_distance_factor
Proportion of vertices, out of the total number of vertices divided by the total number of branches, that can be left at the end of the backbone tree-building algorithm.