Extract the prior information from the milestone network

For example, what are the start cells, the end cells, to which milestone does each cell belong to.

generate_prior_information(cell_ids, milestone_ids, milestone_network,
  milestone_percentages, progressions, divergence_regions, counts,
  feature_info = NULL, cell_info = NULL, marker_logfc = 1,
  marker_minpct = 0.4)

Arguments

cell_ids
The ids of the cells.

milestone_ids
The ids of the milestones in the trajectory. Type: Character vector.

milestone_network
The network of the milestones. Type: Data frame(from = character, to = character, length = numeric, directed = logical).

milestone_percentages
A data frame specifying what percentage milestone each cell consists of. Type: Data frame(cell_id = character, milestone_id = character, percentage = numeric).

progressions
Specifies the progression of a cell along a transition in the milestone_network. Type: Data frame(cell_id = character, from = character, to = character, percentage = numeric).

divergence_regions
A data frame specifying the divergence regions between milestones (e.g. a bifurcation). Type: Data frame(divergence_id = character, milestone_id = character, is_start = logical).

counts
The counts.

feature_info
Optional meta-information pertaining the features.

cell_info
Optional meta-information pertaining the cells.

marker_logfc
Marker genes require at least a X-fold log difference between groups of cells.

marker_minpct
Only test genes that are detected in a minimum fraction of cells between groups of cells.