Inferring trajectories with SCORPIUS

Arguments passed to this function will be used as default parameters for the method.

description_scorpius(distance_method = "spearman", ndim = 3L, k = 4L,
  thresh = 0.001, maxit = 10L, stretch = 0, smoother = "smooth.spline",
  sparse = FALSE)

description_scorpius_sparse(distance_method = "spearman", ndim = 3L,
  k = 4L, thresh = 0.001, maxit = 10L, stretch = 0,
  smoother = "smooth.spline", sparse = TRUE)

Arguments

ndim
The number of dimensions in the new space.

k
The number of clusters to cluster the data into.

thresh
principal.curve parameter: convergence threshhold on shortest distances to the curve

maxit
principal.curve parameter: maximum number of iterations

stretch
principal.curve parameter: a factor by which the curve can be extrapolated when points are projected

smoother
principal.curve parameter: choice of smoother

sparse
Whether or not to use sparse MDS dimensionality reduction, for datasets with large amounts of cells.