Inferring a trajectory inference using CellRouter

Will generate a trajectory using CellRouter.

This method was wrapped inside a container. The original code of this method is available here.

ti_cellrouter(ndim_pca = 20L, ndim_tsne = 11L, max_iter = 1000L,
  cluster_method = "graph.clustering", k_clustering = 20L,
  ndim_pca_clustering = 20L, k_knn = 10L, ndim_pca_knn = 20L,
  sim_type = "jaccard", distance_method_paths = "graph",
  ranks = "rank", num_cells = 3L, neighs = 3L, perplexity = 30L,
  run_environment = NULL)

Arguments

ndim_pca
integer; Number of principlal components to compute (default: 20L; range: from 2L to 100L)

ndim_tsne
integer; Number of tsne dimensions to compute (default: 11L; range: from 2L to 100L)

max_iter
integer; Maximal number of tsne iterations (default: 1000L; range: from 100L to 100000L)

cluster_method
discrete; Method to use for clustering (default: "graph.clustering"; values: "graph.clustering", "model.clustering")

k_clustering
integer; Number of nearest neighbors to build a k-nearest neighbors graph for clustering (default: 20L; range: from 2L to 1000L)

ndim_pca_clustering
integer; Number of PCA dimensions used for k-nearest neighbors graph for clustering (default: 20L; range: from 2L to 100L)

k_knn
integer; Number of nearest neighbors to build a k-nearest neighbors graph for knn (default: 10L; range: from 2L to 1000L)

ndim_pca_knn
integer; Number of PCA dimensions used for knn (default: 20L; range: from 2L to 100L)

sim_type
discrete; Similarity type for knn (default: "jaccard"; values: "jaccard")

distance_method_paths
discrete; Distance method for paths (default: "graph"; values: "euclidean", "maximum", "manhattan", "canberra", "binary", "graph")

ranks
discrete; How to rank the paths (default: "rank"; values: "path_cost", "path_flow", "rank", "length")

num_cells
integer; Trajectories should contain at least num.cells (default: 3L; range: from 3L to 100L)

neighs
integer; The size of the neighborhood in kNN graph used to smoothen kinetic profiles (default: 3L; range: from 2L to 100L)

perplexity
numeric; Perplexity parameter for tsne (default: 30L; range: from 5L to 100L)

run_environment
In which environment to run the method, can be "docker" or "singularity".

Value

A TI method wrapper to be used together with infer_trajectory

References

Lummertz da Rocha, E., Rowe, R.G., Lundin, V., Malleshaiah, M., Jha, D.K., Rambo, C.R., Li, H., North, T.E., Collins, J.J., Daley, G.Q., 2018. Reconstruction of complex single-cell trajectories using CellRouter. Nature Communications 9.