Inferring a trajectory inference using CellTrails

Will generate a trajectory using CellTrails.

This method was wrapped inside a container. The original code of this method is available here.

ti_celltrails(threshold_dl = 2L, threshold_cov = 0.05,
  threshold_ff = 1L, frac = 100L, min_size = 0.01, min_feat = 5L,
  max_pval = 1e-04, min_fc = 2L, l = 10L, run_environment = NULL)

Arguments

threshold_dl
integer; Minimum number of samples; if value < 1 it is interpreted as fraction, otherwise as absolute sample count (default: 2L; range: from 0L to 100L)

threshold_cov
numeric; Minimum coefficient of variation; numeric value between 0 and 1 (default: 0.05; range: from 0L to 1L)

threshold_ff
numeric; A Z-score cutoff (default: 1L; range: from 0L to 5L)

frac
numeric; Fraction or number (if frac > 1) of eigengaps used to perform linear fit. (default: 100L; range: from 1L to 1000L)

min_size
numeric; The initial cluster dedrogram is cut at an height such that the minimum cluster size is at least min_size; if min_size < 1 than the fraction of total samples is used, otherwise it is used as absoulte count (default: 0.01; range: from 0.001 to 1L)

min_feat
integer; Minimum number of differentially expressed features between siblings. If this number is not reached, two neighboring clusters (siblings) in the pruned dendrogram get joined. (default: 5L; range: from 1L to 100L)

max_pval
numeric; Maximum P-value for differential expression computation. (default: 1e-04; range: from 1e-07 to 1L)

min_fc
numeric; Mimimum fold-change for differential expression computation (default: 2L; range: from 1L to 5L)

l
integer; Neighborhood size (default: 10L; range: from 1L to 50L)

run_environment
In which environment to run the method, can be "docker" or "singularity".

Value

A TI method wrapper to be used together with infer_trajectory

References

Ellwanger, D.C., Scheibinger, M., Dumont, R.A., Barr-Gillespie, P.G., Heller, S., 2018. Transcriptional Dynamics of Hair-Bundle Morphogenesis Revealed with CellTrails. Cell Reports 23, 2901–2914.e14.