Inferring trajectories with Embeddr

Inferring trajectories with Embeddr

Will generate a trajectory using Embeddr. This method was wrapped inside a container.

ti_embeddr(docker = "stevedore" %in% rownames(installed.packages()),
  ndim = 2, kernel = "nn", metric = "correlation", nn_pct = 0,
  eps = 0, t = 0, symmetrize = "mean", measure_type = "unorm",
  thresh = 0.001, maxit = 10, stretch = 2, smoother = "smooth.spline")

ti_embeddr(docker = "stevedore" %in% rownames(installed.packages()),
  ndim = 2, kernel = "nn", metric = "correlation", nn_pct = 0,
  eps = 0, t = 0, symmetrize = "mean", measure_type = "unorm",
  thresh = 0.001, maxit = 10, stretch = 2, smoother = "smooth.spline")

Arguments

docker
Whether to use the docker container or the R wrapper

ndim
Dimension of the embedded space, default is 2

kernel
The choice of kernel. 'nn' will give nearest neighbours, 'dist' gives minimum distance and 'heat' gives a heat kernel. Discussed in detail in 'Laplacian Eigenmaps and Spectral Techniques for Embedding and Clustering', Belkin & Niyogi

metric
The metric with which to assess 'closeness' for nearest neighbour selection, one of 'correlation' (pearson) or 'euclidean'. Default is 'correlation'.

nn_pct
The percentage of cells to use as tge number of nearest neighbours if kernel == 'nn'.

eps
Maximum distance parameter if kernel == 'dist'

t
'time' for heat kernel if kernel == 'heat'

symmetrize
How to make the adjacency matrix symmetric. Note that slightly counterintuitively, node i having node j as a nearest neighbour doesn't guarantee node j has node i. There are several ways to get round this:

mean If the above case occurs make the link weight 0.5 so the adjacency matrix becomes \(0.5(A + A')\)

ceil If the above case occurs set the link weight to 1 (ie take the ceiling of the mean case)

floor If the above case occurs set the link weight to 0 (ie take the floor of the mean case)

measure_type
Type of laplacian eigenmap, which corresponds to the constraint on the eigenvalue problem. If type is 'unorm' (default), then the graph measure used is the identity matrix, while if type is 'norm' then the measure used is the degree matrix.

thresh
convergence threshold on shortest distances to the curve.

maxit
maximum number of iterations.

stretch
a factor by which the curve can be extrapolated when points are projected. Default is 2 (times the last segment length). The default is 0 for smoother equal to "periodic_lowess".

smoother
choice of smoother. The default is "smooth_spline", and other choices are "lowess" and "periodic_lowess". The latter allows one to fit closed curves. Beware, you may want to use iter = 0 with lowess().

Value

The trajectory model

Details

This methods was first wrapped inside R, see ti_embeddr

The original code of this method is available here.

The method is described in: Campbell, K., Ponting, C.P., Webber, C., 2015. Laplacian eigenmaps and principal curves for high resolution pseudotemporal ordering of single-cell RNA-seq profiles.