Inferring a trajectory inference using FateID

Will generate a trajectory using FateID. This method was wrapped inside a container.

ti_fateid(reclassify = TRUE, clthr = 0.9, nbfactor = 5L, q = 0.75,
  k = 3L, m = "tsne", minnr = 5L, minnrh = 10L, trthr = 0.4)

Arguments

reclassify
Whether to reclassify the cell grouping

clthr
Real number between zero and one. This is the threshold for the fraction of random forest votes required to assign a cell not contained within the target clusters to one of these clusters. The value of this parameter should be sufficiently high to only reclassify cells with a high-confidence assignment. Default value is 0.9.
numeric; default: 0.9; possible values between 0.1 and 1

nbfactor
Positive integer number. Determines the number of trees grown for each random forest. The number of trees is given by the number of columns of th training set multiplied by nbfactor. Default value is 5.
integer; default: 5L; possible values between 2 and 100

q
Q real value between zero and one. This number specifies a threshold used for feature selection based on importance sampling. A reduced expression table is generated containing only features with an importance larger than the q-quantile for at least one of the classes (i. e. target clusters). Default value is 0.75.
numeric; default: 0.75; possible values between 0 and 1

k
Number of dimensions
integer; default: 3L; possible values between 2 and 100

m
Dimensionality reduction method to use. Can be tsne, cmd, dm or lle
discrete; default: "tsne"; possible values: tsne, cmd, dm, lle

minnr
Integer number of cells per target cluster to be selected for classification (test set) in each round of training. For each target cluster, the minnr cells with the highest similarity to a cell in the training set are selected for classification. If z is not NULL it is used as the similarity matrix for this step. Otherwise, 1-cor(x) is used. Default value is 5.
integer; default: 5L; possible values between 2 and 100

minnrh
Integer number of cells from the training set used for classification. From each training set, the minnrh cells with the highest similarity to the training set are selected. If z is not NULL it is used as the similarity matrix for this step. Default value is 10.
integer; default: 10L; possible values between 2 and 100

trthr
Real value representing the threshold of the fraction of random forest votes required for the inclusion of a given cell for the computation of the principal curve. If NULL then only cells with a significant bias >1 are included for each trajectory. The bias is computed as the ratio of the number of votes for a trajectory and the number of votes for the trajectory with the second largest number of votes. By this means only the trajectory with the largest number of votes will receive a bias >1. The siginifcance is computed based on counting statistics on the difference in the number of votes. A significant bias requires a p-value < 0.05.
numeric; default: 0.4; possible values between 0 and 1

Value

The trajectory model

Details

The original code of this method is available here.

The method is described in: Herman, J.S., Sagar, Grün, D., 2018. FateID infers cell fate bias in multipotent progenitors from single-cell RNA-seq data. Nature Methods 15, 379–386.