Inferring a trajectory inference using MERLoT

Will generate a trajectory using MERLoT. This method was wrapped inside a container.

ti_merlot(sigma = "local", distance = "euclidean", ndim = 20L,
  density_norm = TRUE, n_local_lower = 5L, n_local_upper = 7L,
  w_width = 0.01, n_components_to_use = 3, N_yk = 100, lambda_0 = 1e-10,
  mu_0 = 0.0025, increaseFactor_mu = 20, increaseFactor_lambda = 20,
  FixEndpoints = FALSE)

Arguments

sigma
Diffusion scale parameter of the Gaussian kernel. A larger sigma might be necessary if the eigenvalues can not be found because of a singularity in the matrix. Must be one of:

a numeric global sigma -- a global sigma will be calculated using destiny::find_sigmas()

or a destiny::Sigmas-class() object.

discrete; default: "local"; possible values: local, global

distance
A stats::dist() object, or a character vector specifying which distance metric to use. Allowed measures:

cosine distance (1-corr(c_1, c_2)), or

rank correlation distance (1-corr(rank(c_1), rank(c_2)))

discrete; default: "euclidean"; possible values: euclidean, cosine, rankcor

ndim
Number of eigenvectors/dimensions to return
integer; default: 20L; possible values between 2 and 20

density_norm
Logical. If TRUE, use density normalisation
logical; default: TRUE; possible values: TRUE, FALSE

n_local_lower
If sigma == 'local', the n_local_lower:n_local_upper nearest neighbor(s) determine(s) the local sigma
integer; default: 5L; possible values between 2 and 20

n_local_upper
See n_local_lower
integer; default: 7L; possible values between 2 and 20

w_width
Window width to use for deciding the branch cutoff
numeric; default: 0.01; possible values between 1e-04 and 1

n_components_to_use
Which components to use in downstream analysis
integer; default: 3; possible values between 2 and 20

N_yk
Number of nodes for the elastic principal tree
integer; default: 100; possible values between 2 and 1000

lambda_0
Principal elastic tree energy function parameter.
numeric; default: 1e-10; possible values between 1e-15 and 1e-04

mu_0
Principal elastic tree energy function parameter.
numeric; default: 0.0025; possible values between 5e-04 and 0.005

increaseFactor_mu
Factor by which the mu will be increased for the embedding
numeric; default: 20; possible values between 2 and 50

increaseFactor_lambda
Factor by which the mu will be increased for the embedding
numeric; default: 20; possible values between 2 and 50

FixEndpoints
Documentation not provided by authors
logical; default: FALSE; possible values: TRUE, FALSE

Value

The trajectory model

Details

The original code of this method is available here.

The method is described in: Parra, R.G., Papadopoulos, N., Ahumada-Arranz, L., El Kholtei, J., Treutlein, B., Soeding, J., 2018. Reconstructing complex lineage trees from scRNA-seq data using MERLoT.