Inferring a trajectory inference using mfa

Will generate a trajectory using mfa.

This method was wrapped inside a container. The original code of this method is available here.

ti_mfa(b = 2L, iter = 2000L, thin = 1L, pc_initialise = 1L,
  prop_collapse = 0, scale_input = TRUE, zero_inflation = FALSE,
  run_environment = NULL)

Arguments

b
integer; Number of branches to model (default: 2L; range: from 1L to 10L)

iter
integer; Number of MCMC iterations (default: 2000L; range: from 20L to 5000L)

thin
integer; MCMC samples to thin (default: 1L; range: from 1L to 20L)

pc_initialise
integer; Which principal component to initialise pseudotimes to (default: 1L; range: from 1L to 5L)

prop_collapse
numeric; Proportion of Gibbs samples which should marginalise over c (default: 0; range: from 0 to 1)

scale_input
logical; Logical. If true, input is scaled to have mean 0 variance 1

zero_inflation
logical; Logical, should zero inflation be enabled?

run_environment
In which environment to run the method, can be "docker" or "singularity".

Value

A TI method wrapper to be used together with infer_trajectory

References

Campbell, K.R., Yau, C., 2017. Probabilistic modeling of bifurcations in single-cell gene expression data using a Bayesian mixture of factor analyzers. Wellcome Open Research 2, 19.