Inferring a trajectory inference using paga

Will generate a trajectory using paga. This method was wrapped inside a container.

ti_paga(n_neighbors = 30L, n_comps = 50L, resolution = 2.5)

Arguments

n_neighbors
Number of neighbours for knn
integer; default: 30L; possible values between 1 and 100

n_comps
Number of principal components
integer; default: 50L; possible values between 0 and 100

resolution
Resolution of louvain clustering, which determines the granularity of the clustering. Higher values will result in more clusters.
numeric; default: 2.5; possible values between 0.1 and 10

Value

The trajectory model

Details

The original code of this method is available here.

The method is described in: Wolf, F.A., Hamey, F., Plass, M., Solana, J., Dahlin, J.S., Gottgens, B., Rajewsky, N., Simon, L., Theis, F.J., 2017. Graph abstraction reconciles clustering with trajectory inference through a topology preserving map of single cells.