Inferring a trajectory inference using Projected Monocle

Will generate a trajectory using Projected Monocle.

This method was wrapped inside a container. The original code of this method is available here.

ti_projected_monocle(reduction_method = "DDRTree", max_components = 2L,
  norm_method = "vstExprs", auto_param_selection = TRUE,
  run_environment = NULL)

Arguments

reduction_method
discrete; A character string specifying the algorithm to use for dimensionality reduction. (default: "DDRTree"; values: "DDRTree")

max_components
integer; The dimensionality of the reduced space (default: 2L; range: from 2L to 20L)

norm_method
discrete; Determines how to transform expression values prior to reducing dimensionality (default: "vstExprs"; values: "vstExprs", "log", "none")

auto_param_selection
logical; When this argument is set to TRUE (default), it will automatically calculate the proper value for the ncenter (number of centroids) parameters which will be passed into DDRTree call.

run_environment
In which environment to run the method, can be "docker" or "singularity".

Value

A TI method wrapper to be used together with infer_trajectory