Inferring a trajectory inference using pseudogp

Will generate a trajectory using pseudogp.

This method was wrapped inside a container. The original code of this method is available here.

ti_pseudogp(smoothing_alpha = 10, smoothing_beta = 3,
  pseudotime_mean = 0.5, pseudotime_var = 1, chains = 3L,
  iter = 100L, dimreds = c(TRUE, TRUE, FALSE, FALSE, FALSE, FALSE,
  FALSE, FALSE, FALSE, FALSE, FALSE), initialise_from = "random",
  run_environment = NULL)

Arguments

smoothing_alpha
numeric; The hyperparameter for the Gamma distribution that controls arc-length (default: 10; range: from 1 to 20)

smoothing_beta
numeric; The hyperparameter for the Gamma distribution that controls arc-length (default: 3; range: from 1 to 20)

pseudotime_mean
numeric; The mean of the constrained normal prior on the pseudotimes (default: 0.5; range: from 0 to 1)

pseudotime_var
numeric; The variance of the constrained normal prior on the pseudotimes (default: 1; range: from 0.01 to 1)

chains
integer; The number of chains for the MCMC trace (default: 3L; range: from 1L to 20L)

iter
integer; The number of iterations for the MCMC trace (default: 100L; range: from 100L to 1000L)

dimreds
logical_vector; A character vector specifying which dimensionality reduction methods to use.See dyndimred::dimred for the list of available dimensionality reduction methods.

initialise_from
discrete; How to initialise the MCMC chain. One of "random" (stan decides),"principal_curve", or "pca" (the first component of PCA rescaled is taken to be the pseudotimes).Note: if multiple representations are provided, pseudogp will take the principal curve orpca from the first rather than combining them. If a particular representation is required, it isup to the user to re-order them. (default: "random"; values: "random", "principal_curve", "pca")

run_environment
In which environment to run the method, can be "docker" or "singularity".

Value

A TI method wrapper to be used together with infer_trajectory

References

Campbell, K.R., Yau, C., 2016. Order Under Uncertainty: Robust Differential Expression Analysis Using Probabilistic Models for Pseudotime Inference. PLOS Computational Biology 12, e1005212.