Inferring a trajectory inference using SCORPIUS sparse

Will generate a trajectory using SCORPIUS sparse. This method was wrapped inside a container.

ti_scorpius_sparse(docker = dynwrap::test_docker_installation(),
  distance_method = "spearman", ndim = 3, k = 4, thresh = 0.001,
  maxit = 10, stretch = 0, smoother = "smooth_spline", sparse = TRUE)

ti_scorpius_sparse(docker = dynwrap::test_docker_installation(),
  distance_method = "spearman", ndim = 3, k = 4, thresh = 0.001,
  maxit = 10, stretch = 0, smoother = "smooth_spline", sparse = TRUE)

Arguments

docker
Whether to use the docker container or the R wrapper

distance_method
A character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson", "kendall", or "spearman".

ndim
The number of dimensions in the new space.

k
The number of clusters to cluster the data into.

thresh
principal.curve parameter: convergence threshhold on shortest distances to the curve

maxit
principal.curve parameter: maximum number of iterations

stretch
principal.curve parameter: a factor by which the curve can be extrapolated when points are projected

smoother
principal.curve parameter: choice of smoother

sparse
Whether or not to use sparse MDS dimensionality reduction, for datasets with large amounts of cells.

Value

The trajectory model

Details

This methods was first wrapped inside R, see ti_scorpius_sparse

The original code of this method is available here.

The method is described in: Cannoodt, R., Saelens, W., Sichien, D., Tavernier, S., Janssens, S., Guilliams, M., Lambrecht, B.N., De Preter, K., Saeys, Y., 2016. SCORPIUS improves trajectory inference and identifies novel modules in dendritic cell development.