Inferring a trajectory inference using SCOUP

Will generate a trajectory using SCOUP.

This method was wrapped inside a container. The original code of this method is available here.

ti_scoup(ndim = 2L, max_ite1 = 100, max_ite2 = 100,
  alpha_min = 0.1, alpha_max = 100, t_min = 0.001, t_max = 2,
  sigma_squared_min = 0.1, thresh = 0.01, run_environment = NULL)

Arguments

ndim
integer; Number of pca dimensions (default: 2L; range: from 2L to 20L)

max_ite1
numeric; Upper bound of EM iteration (without pseudo-time optimization). The detailed explanation is described in the supplementary text. (default is 1,000) (default: 100; range: from 2 to 5000)

max_ite2
numeric; Upper bound of EM iteration (including pseudo-time optimization) (default is 1,000). (default: 100; range: from 2 to 5e+05)

alpha_min
numeric; Lower bound of alpha (default is 0.1) (default: 0.1; range: from 0.001 to 10)

alpha_max
numeric; Upper bound of alpha (default is 100) (default: 100; range: from 1 to 10000)

t_min
numeric; Lower bound of pseudo-time (default is 0.001) (default: 0.001; range: from 1e-05 to 1)

t_max
numeric; Upper bound of pseudo-time (default is 2.0) (default: 2; range: from 0.1 to 100)

sigma_squared_min
numeric; Lower bound of sigma squared (default is 0.1) (default: 0.1; range: from 0.001 to 10)

thresh
numeric; Threshold (default: 0.01; range: from 0.01 to 10)

run_environment
In which environment to run the method, can be "docker" or "singularity".

Value

A TI method wrapper to be used together with infer_trajectory

References

Matsumoto, H., Kiryu, H., 2016. SCOUP: a probabilistic model based on the Ornstein–Uhlenbeck process to analyze single-cell expression data during differentiation. BMC Bioinformatics 17.