Inferring a trajectory inference using SCUBA

Will generate a trajectory using SCUBA.

This method was wrapped inside a container. The original code of this method is available here.

ti_scuba(rigorous_gap_stats = TRUE, N_dim = 2L,
  low_gene_threshold = 1L, low_gene_fraction_max = 0.7,
  min_split = 15L, min_percentage_split = 0.25,
  run_environment = NULL)

Arguments

rigorous_gap_stats
logical; Whether to use rigorous gap statistics to determine number of clusters

N_dim
integer; Number of TSNE dimensions (default: 2L; range: from 2L to 3L)

low_gene_threshold
numeric; Threshold value for genes of low expression levels (default: 1L; range: from 0L to 5L)

low_gene_fraction_max
numeric; Maximum fraction of lowly-expressed cells allowed for each gene (default: 0.7; range: from 0L to 1L)

min_split
integer; Lower threshold on the number of cells in a cluster for this cluster to be split. (default: 15L; range: from 1L to 100L)

min_percentage_split
numeric; Minimum fraction of cells in the smaller cluster during a bifurcation. (default: 0.25; range: from 0L to 1L)

run_environment
In which environment to run the method, can be "docker" or "singularity".

Value

A TI method wrapper to be used together with infer_trajectory

References

Marco, E., Karp, R.L., Guo, G., Robson, P., Hart, A.H., Trippa, L., Yuan, G.-C., 2014. Bifurcation analysis of single-cell gene expression data reveals epigenetic landscape. Proceedings of the National Academy of Sciences 111, E5643–E5650.