Inferring trajectories with Slingshot

Inferring trajectories with Slingshot

Will generate a trajectory using Slingshot. This method was wrapped inside a container.

ti_slingshot(docker = dynwrap::test_docker_installation(), ndim = 3,
  nclus = 5, dimred = "pca", shrink = 1, reweight = TRUE, thresh = -3,
  maxit = 10, stretch = 2, smoother = "smooth.spline",
  shrink.method = "cosine")

ti_slingshot(docker = dynwrap::test_docker_installation(), ndim = 3,
  nclus = 5, dimred = "pca", shrink = 1, reweight = TRUE, thresh = -3,
  maxit = 10, stretch = 2, smoother = "smooth.spline",
  shrink.method = "cosine")

Arguments

docker
Whether to use the docker container or the R wrapper

ndim
The number of dimensions

nclus
Number of clusters

dimred
A character vector specifying which dimensionality reduction method to use. See dyndimred::dimred for the list of available dimensionality reduction methods.

shrink
logical or numeric between 0 and 1, determines whether and how much to shrink branching lineages toward their average prior to the split.

reweight
logical, whether to allow cells shared between lineages to be reweighted during curve-fitting. If TRUE, cells shared between lineages will be iteratively reweighted based on the quantiles of their projection distances to each curve. See 'Details' for more.

thresh
numeric, determines the convergence criterion. Percent change in the total distance from cells to their projections along curves must be less than thresh. Default is 0.001, similar to principal.curve.

maxit
numeric, maximum number of iterations, see principal.curve.

stretch
numeric factor by which curves can be extrapolated beyond endpoints. Default is 2, see principal.curve.

smoother
choice of scatter plot smoother. Same as principal.curve, but "lowess" option is replaced with "loess" for additional flexibility.

shrink.method
character denoting how to determine the appropriate amount of shrinkage for a branching lineage. Accepted values are the same as for kernel in density (default is "cosine"), as well as "tricube" and "density". See 'Details' for more.

Value

The trajectory model

Details

This methods was first wrapped inside R, see ti_slingshot

The original code of this method is available here.

The method is described in: Street, K., Risso, D., Fletcher, R.B., Das, D., Ngai, J., Yosef, N., Purdom, E., Dudoit, S., 2017. Slingshot: Cell lineage and pseudotime inference for single-cell transcriptomics.

See also