Inferring trajectories with StemID

Inferring trajectories with StemID

Will generate a trajectory using StemID. This method was wrapped inside a container.

ti_stemid(docker = dynwrap::test_docker_installation(), clustnr = 30,
  bootnr = 50, metric = "pearson", num_cluster_method = "sat",
  SE.method = "Tibs2001SEmax", SE.factor = 0.25, B.gap = 50, cln = 30,
  FUNcluster = "kmedoids", dimred_method = "tsne", outminc = 0,
  outlg = 2, probthr = 0.001, thr_lower = -10, thr_upper = -5,
  outdistquant = 0.95, nmode = FALSE, pdishuf = 2000, pthr = 1e-04,
  pethr = 1e-04, pvalue_cutoff = 0.05, linkscore_cutoff = 0.2)

ti_stemid(docker = dynwrap::test_docker_installation(), clustnr = 30,
  bootnr = 50, metric = "pearson", num_cluster_method = "sat",
  SE.method = "Tibs2001SEmax", SE.factor = 0.25, B.gap = 50, cln = 30,
  FUNcluster = "kmedoids", dimred_method = "tsne", outminc = 0,
  outlg = 2, probthr = 0.001, thr_lower = -10, thr_upper = -5,
  outdistquant = 0.95, nmode = FALSE, pdishuf = 2000, pthr = 1e-04,
  pethr = 1e-04, pvalue_cutoff = 0.05, linkscore_cutoff = 0.2)

Arguments

docker
Whether to use the docker container or the R wrapper

clustnr
maximum number of clusters for the computation of the gap statistic or derivation of the cluster number by the saturation criterion. Default is 30. If more major cell types are expected a higher number should be chosen.

bootnr
number of booststrapping runs for clusterboot. Default is 50

metric
the input data are transformed to a distance object. Distances can be computed based on different metrics. Possible values are "pearson", "spearman", "logpearson", "euclidean", "kendall", "maximum", "manhattan", "canberra", "binary" or "minkowski". Default is "pearson". In case of the correlation based methods, the distance is computed as 1 - correlation. K-medoids clustering is performed on this distance object.

num_cluster_method
the type of clustering method, can be sat, gap or manual

SE.method
the clustering routine calls a modified version of the maxSE function from the cluster package to determine the first local maximum of the gap statistic. By default, we use the method "Tibs2001SEmax" for calling the first local maximum (see specification of maxSE). This method requires that the maximum exceeds the values of its neighbors by a fraction of their standard deviation. This fraction is defined by the parameter SE.factor. All methods defined for the original maxSE function can also be used.

SE.factor
fraction of the standard deviation by which the local maximum is required to differ from the neighboring points it is compared to. Default is 0.25.

B.gap
number of bootstrap runs for the calculation of the gap statistic. Default is 50.

cln
the number of clusters for k-medoids clustering. Default is 0. In this case, the cluster number is determined based on the gap statistic6 and do.gap has to be TRUE.

FUNcluster
the clustering method applied. One of the following methods can be selected: kmedoids, kmeans, hclust. RaceID3 is designed for k-medoids clustering and therefore it is recommended to use only the kmedoids method. Default is kmedoids.

dimred_method
the dimensionality reduction method, can be tsne, sammon or tsne_initcmd

outminc
expression cutoff for the identification of outlier genes is defined. Default is 5.

outlg
minimal number of outlier genes required to identify a cell as an outlier. Default is 2.

probthr
defines the probability threshold for outlier calling. If the probability of observing a given expression level for a gene in a cell is lower than this cutoff (based on the negative binomial distribution for the calibrated noise model), the cell is considered an outlier for this gene. Default is 10-3.

thr_lower
lower probability for which the number of outliers is computed in order to plot the dependence of the number of outliers on the probability threshold

thr_upper
upper probability for which the number of outliers is computed in order to plot the dependence of the number of outliers on the probability threshold

outdistquant
outlier cells are merged to outlier clusters if their similarity exceeds the outdistquant-quantile of the similarity distribution for all pairs of cells that are together in one of the original clusters. Default is 0.95.

nmode
Boolean argument. If nmode is set to TRUE the assignment to inter-cluster links for each cell is not done based on the longest projection, but based on identifying the cluster (other than the cluster the cell belongs to) that contains the nearest neighbor of the cell, i. e. the cell with the most similar transcriptome. The coordinate on the assigned link is still derived based on the projection. Default is FALSE.

pdishuf
positive integer. This is the number of randomizations to be performed. As a rule of thumb this number should be at least one order of magnitude larger than the desired p-value on the significance of the number of cells on a connection. Default is 2000.

pthr
positive number. This number corresponds to the p-value threshold, which is used to determine, whether the magnitude of an observed trajectory is significantly larger than observed for the randomized background distribution. This criterion is not used to infer significance of a link, but shown in a graphical representation of the tree

pethr
positive number. This number corresponds to the p-value threshold, which isused to determine for each link if it is populated by a number of cellssignificantly larger than expected for the randomized background distribution. This p-value threshold determines, which connections are considered validdifferentiation trajectories in the derived lineage tree.

pvalue_cutoff
Upper cutoff of p-value for determining significant edges

linkscore_cutoff
Lower cutoff of the linkscore for determining significant edges

Value

The trajectory model

Details

This methods was first wrapped inside R, see ti_stemid

The original code of this method is available here.

The method is described in: Grün, D., Muraro, M.J., Boisset, J.-C., Wiebrands, K., Lyubimova, A., Dharmadhikari, G., van den Born, M., van Es, J., Jansen, E., Clevers, H., de Koning, E.J.P., van Oudenaarden, A., 2016. De Novo Prediction of Stem Cell Identity using Single-Cell Transcriptome Data. Cell Stem Cell 19, 266–277.