Inferring a trajectory inference using URD

Will generate a trajectory using URD.

This method was wrapped inside a container. The original code of this method is available here.

ti_urd(knn = 0L, sigma.use = 0, distance = "euclidean",
  n_floods = 20L, stability.div = 10L, mp.factor = 1L,
  perplexity = 30L, theta = 0.5, max_iter = 1000L, num.nn = 30L,
  do.jaccard = TRUE, optimal.cells.forward = 20L,
  max.cells.back = 40L, n.per.tip = 25000L, root.visits = 1L,
  max.steps = 25000L, n.subsample = 10L, divergence.method = "ks",
  cells.per.pseudotime.bin = 80L, bins.per.pseudotime.window = 5L,
  p.thresh = 0.01, run_environment = NULL)

Arguments

knn
integer; Number of nearest neighbors to use. 0 takes a guess. (default: 0L; range: from 0L to 50L)

sigma.use
numeric; Kernel width to use for the diffusion map. 0 uses destiny's global auto-detection procedure. (default: 0; range: from 0 to 1)

distance
discrete; Distance metric to use for determining transition probabilities. (default: "euclidean"; values: "euclidean", "cosine", "rankcor")

n_floods
integer; Number of simulations to perform and average. (default: 20L; range: from 5L to 50L)

stability.div
numeric; Number of simulation subsamplings to calculate. (default: 10L; range: from 2L to 50L)

mp.factor
numeric; Retain PCs than are this factor more than the estimated maximum singular value expected or random data. This is useful in cases when there are many PCs that have standard deviations just above that expected by random, which probably represent noise and should be excluded. (default: 1L; range: from 0L to 10L)

perplexity
numeric; Perplexity parameter for the tSNE. (default: 30L; range: from 0L to 100L)

theta
numeric; Speed/accuracy trade-off for Barnes-Hut approximation of tSNE. 0 is exact tSNE, higher is less accurate. (default: 0.5; range: from 0L to 1L)

max_iter
integer; Number of nearest neighbors to use. 0 takes a guess. (default: 1000L; range: from 100L to 10000L)

num.nn
integer; How many nearest-neighbors to use in the k-nn graph. (default: 30L; range: from 10L to 100L)

do.jaccard
logical; Weight edges in the k-nn graph according to their Jaccard overlap?

optimal.cells.forward
numeric; The number of cells in the direction specified by pseudotime.direction at which the logistic should reach 1-asymptote. (default: 20L; range: from 5L to 100L)

max.cells.back
numeric; The number of cells in the direction opposite from that specified by pseudotime.direction at which the logistic should reach asymptote. (default: 40L; range: from 5L to 200L)

n.per.tip
integer; Number of walks to do per tip. (default: 25000L; range: from 100L to 1000000L)

root.visits
integer; Number of steps to take that visit a root.cell before stopping. (default: 1L; range: from 1L to 5L)

max.steps
integer; Number of walks to do per tip. (default: 25000L; range: from 100L to 1000000L)

n.subsample
integer; Number of subsamplings to perform for calculating stability. (default: 10L; range: from 2L to 100L)

divergence.method
discrete; Distance metric to use for determining transition probabilities. (default: "ks"; values: "ks", "preference")

cells.per.pseudotime.bin
integer; Approximate number of cells to assign to each pseudotime bin for branchpoint finding. (default: 80L; range: from 10L to 1000L)

bins.per.pseudotime.window
integer; Width of moving window in pseudotime used for branchpoint finding, in terms of bins. (default: 5L; range: from 2L to 20L)

p.thresh
numeric; P-value threshold to use in determining whether visitation is significantly different from pairs of tips (default: 0.01; range: from 1e-05 to 1L)

run_environment
In which environment to run the method, can be "docker" or "singularity".

Value

A TI method wrapper to be used together with infer_trajectory

References

Farrell, J.A., Wang, Y., Riesenfeld, S.J., Shekhar, K., Regev, A., Schier, A.F., 2018. Single-cell reconstruction of developmental trajectories during zebrafish embryogenesis. Science 360, eaar3131.