Inferring trajectories with MERLoT

Inferring trajectories with MERLoT

ti_merlot(sigma = "local", distance = "euclidean", ndim = 20L,
  density_norm = TRUE, n_local_lower = 5L, n_local_upper = 7L,
  w_width = 0.1, n_components_to_use = 3L, N_yk = 100L,
  lambda_0 = 4.53999297624849e-05, mu_0 = 0.0025, increaseFactor_mu = 20,
  increaseFactor_lambda = 20, FixEndpoints = FALSE)

Arguments

sigma
Diffusion scale parameter of the Gaussian kernel. A larger sigma might be necessary if the eigenvalues can not be found because of a singularity in the matrix. Must be one of:

A character vector: "local" (default) or "global",

a numeric global sigma -- a global sigma will be calculated using find_sigmas

or a Sigmas-class object.

distance
A dist object, or a character vector specifying which distance metric to use. Allowed measures:

Euclidean distance (default),

cosine distance (1-corr(c_1, c_2)), or

rank correlation distance (1-corr(rank(c_1), rank(c_2)))

ndim
Number of eigenvectors/dimensions to return

density_norm
logical. If TRUE, use density normalisation

n_local_lower
If sigma == 'local', the n_local_lower:n_local_upper nearest neighbor(s) determine(s) the local sigma

n_local_upper
See n_local_lower

w_width
Window width to use for deciding the branch cutoff

n_components_to_use
Which components to use in downstream analysis

N_yk
number of nodes for the elastic principal tree

lambda_0
principal elastic tree energy function parameter.

mu_0
principal elastic tree energy function parameter.

increaseFactor_mu
factor by which the mu will be increased for the embedding

increaseFactor_lambda
factor by which the mu will be increased for the embedding

FixEndpoints
Documentation not provided by authors