Inferring trajectories with mfa

Inferring trajectories with mfa

Will generate a trajectory using mfa. This method was wrapped inside a container.

ti_mfa(docker = "stevedore" %in% rownames(installed.packages()), b = 2,
  iter = 2000, thin = 1, pc_initialise = 1, prop_collapse = 0,
  scale_input = TRUE, zero_inflation = FALSE)

ti_mfa(docker = "stevedore" %in% rownames(installed.packages()), b = 2,
  iter = 2000, thin = 1, pc_initialise = 1, prop_collapse = 0,
  scale_input = TRUE, zero_inflation = FALSE)

Arguments

docker
Whether to use the docker container or the R wrapper

b
Number of branches to model

iter
Number of MCMC iterations

thin
MCMC samples to thin

pc_initialise
Which principal component to initialise pseudotimes to

prop_collapse
Proportion of Gibbs samples which should marginalise over c

scale_input
Logical. If true, input is scaled to have mean 0 variance 1

zero_inflation
Logical, should zero inflation be enabled?

Value

The trajectory model

Details

This methods was first wrapped inside R, see ti_mfa

The original code of this method is available here.

The method is described in: Campbell, K.R., Yau, C., 2017. Probabilistic modeling of bifurcations in single-cell gene expression data using a Bayesian mixture of factor analyzers. Wellcome Open Research 2, 19.

See also