Inferring trajectories with SLICE

Inferring trajectories with SLICE

Will generate a trajectory using SLICE. This method was wrapped inside a container.

ti_slice(docker = dynwrap::test_docker_installation(),
  lm.method = "clustering", model.type = "tree", ss.method = "all",
  ss.threshold = 0.25, community.method = "louvain",
  cluster.method = "kmeans", k = 0, k.max = 10, B = 100,
  k.opt.method = "firstmax")

ti_slice(docker = dynwrap::test_docker_installation(),
  lm.method = "clustering", model.type = "tree", ss.method = "all",
  ss.threshold = 0.25, community.method = "louvain",
  cluster.method = "kmeans", k = 0, k.max = 10, B = 100,
  k.opt.method = "firstmax")

Arguments

docker
Whether to use the docker container or the R wrapper

lm.method
Select "clustering" based or "graph" based method to infer lineage model

model.type
The type of models that will be infered: "tree" - directed minimum spanning tree based, "graph" - directed graph based

ss.method
The method for defining core cell set for stable state detection: all - all the cells in a cluster constitute the core cell set; top - cells with scEntropy lower than the ss.threshold quantile of all the values in a cluster constitute the core cell set; pcst - cells with scEntropy lower than the ss.threshold quantile of all the values in a cluster constitute the prize nodes, linear prize-collecting steiner tree algorithm is used to approximate an optimal subnetwork, the cells in the subnetwork constitute the core cell set. Stable states are defined as the centroids of the core cell sets.

ss.threshold
The threshold used when ss.method is "top" or "pcst". Default: 0.25.

community.method
The method for network community detection. Most of the community detection methods implemented in the igraph package are supported, including "fast_greedy", "edge_betweenness", "label_prop", "leading_eigen","louvain","spinglass", "walktrap". If this parameter is set to "auto", the algorithm will perform all the community detection methods and select the one that generates the communities with best modularity. Only take effect when lm.method is "graph"

cluster.method
Use "kmeans" or "pam" to divide cells into clusters. Only take effect when lm.method is "clustering"

k
The number of cell clusters. If NULL, Gap statistic will be used to determine an optimal k.

k.max
The "k.max" parameter of cluster::clusGap(); used when k is NULL.

B
The "B" parameter of cluster::clusGap(); used when k is NULL

k.opt.method
The "method" parameter of cluster::maxSE(); used when k is NULL

Value

The trajectory model

Details

This methods was first wrapped inside R, see ti_slice

The original code of this method is available here.

The method is described in: Guo, M., Bao, E.L., Wagner, M., Whitsett, J.A., Xu, Y., 2016. SLICE: determining cell differentiation and lineage based on single cell entropy. Nucleic Acids Research gkw1278.